Refactored the mob spawner and did some minor adjustments to other classes

+++ src/main/java/org/clawd/buttons/type/HitButton.java
@@ -2,6 +2,7 @@ package org.clawd.buttons.type;
 
 import net.dv8tion.jda.api.events.interaction.component.ButtonInteractionEvent;
 import org.clawd.buttons.CustomButton;
+import org.clawd.data.mobs.Mob;
 import org.clawd.main.Main;
 import org.clawd.tokens.Constants;
 
@@ -16,10 +17,18 @@ public class HitButton implements CustomButton {
             String buttonID = event.getComponentId();
             int mobID = Integer.parseInt(buttonID.replace(Constants.HIT_BUTTON_ID, ""));
 
-            // TODO
+            Mob spawnedMob = Main.mineworld.getMobByID(mobID);
 
-            event.deferEdit().queue();
+            if (spawnedMob == null)
+                return;
+
             event.getMessage().delete().queue();
+            event.deferEdit().queue();
+
+            /*
+             * TODO:
+             *  - Implement logic regarding loot: xp, gold, etc.
+             */
 
             Main.LOG.info("Executed '" + Constants.HIT_BUTTON_ID + "' button");
         }
